♻️ abstracting common code
