+++ Sources/ChangeMenuBarColor/Commands/SolidColor.swift
@@ -10,7 +10,7 @@ import Foundation
 import Cocoa
 import SwiftHEXColors
 
-struct SolidColor: ParsableCommand {
+struct SolidColor: Command {
     static let configuration = CommandConfiguration(
         commandName: "SolidColor",
         abstract: "Adds solid color rectangle"
@@ -22,55 +22,36 @@ struct SolidColor: ParsableCommand {
     @Argument(help: "HEX color to use for the menu bar")
     private var color: String
 
-    func run() {
+    func createWallpaper(screenSize: CGSize, menuBarHeight: CGFloat) -> NSImage? {
         guard let color: NSColor = NSColor(hexString: self.color) else {
             print("Invalid HEX color provided. Make sure it includes the '#' symbol, e.g: #FF0000")
-            return
+            return nil
         }
 
         guard let wallpaper = NSImage(contentsOfFile: wallpaper) else {
             print("Cannot read the provided wallpaper file as image")
-            return
+            return nil
         }
 
         guard !NSScreen.screens.isEmpty else {
             print("Cannot detect screens")
-            return
+            return nil
         }
 
-        var index = 0
-
-        for screen in NSScreen.screens {
-            index = index+1
-
-            let screenSize = screen.frame.size
-            let menuBarHeight = screenSize.height - screen.visibleFrame.height - screen.visibleFrame.origin.y
-
-            guard let resizedWallapper = wallpaper.resized(to: screenSize) else {
-                print("Cannot not resize provided wallpaper to screen size")
-                return
-            }
-
-            guard let topImage = createSolidImage(color: color, width: screenSize.width, height: menuBarHeight) else {
-                return
-            }
-
-            guard let combined = combineImages(baseImage: resizedWallapper, addedImage: topImage), let data = combined.jpgData else {
-                return
-            }
-
-            let workingDirectory = FileManager.default.currentDirectoryPath
-            let adjustedWallpaperFile = workingDirectory.appending("/wallpaper-screen\(index)-adjusted.jpg")
-
-            do {
-                try data.write(to: URL(fileURLWithPath: adjustedWallpaperFile))
-                print("Created new wallpaper \(adjustedWallpaperFile)")
-            } catch {
-                print("Writing new wallpaper file failed with \(error.localizedDescription)")
-            }
+        guard let resizedWallapper = wallpaper.resized(to: screenSize) else {
+            print("Cannot not resize provided wallpaper to screen size")
+            return nil
         }
 
-        print("\nAll done!")
+        guard let topImage = createSolidImage(color: color, width: screenSize.width, height: menuBarHeight) else {
+            return nil
+        }
+
+        return combineImages(baseImage: resizedWallapper, addedImage: topImage)
+    }
+
+    func run() {
+        process()
     }
 }
